+++ app/src/main/java/com/techcourse/dao/UserDao.java
@@ -1,11 +1,9 @@
 package com.techcourse.dao;
 
-import com.interface21.jdbc.JdbcTemplate;
+import com.interface21.jdbc.core.JdbcTemplate;
 import com.interface21.jdbc.RowMapper;
 import com.techcourse.domain.User;
-import java.sql.Connection;
 import java.util.List;
-import javax.sql.DataSource;
 
 public class UserDao {
 
@@ -18,8 +16,8 @@ public class UserDao {
 
     private final JdbcTemplate jdbcTemplate;
 
-    public UserDao(final DataSource dataSource) {
-        this.jdbcTemplate = new JdbcTemplate(dataSource);
+    public UserDao(final JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
     }
 
     public void insert(final User user) {
@@ -30,14 +28,6 @@ public class UserDao {
     public void update(final User user) {
         final var sql = "update users set password = ?, email = ? where account = ?";
         jdbcTemplate.update(sql, user.getPassword(), user.getEmail(), user.getAccount());
-    }
-
-    public void update(
-            final Connection connection,
-            final User user
-    ) {
-        final var sql = "update users set password = ?, email = ? where account = ?";
-        jdbcTemplate.update(connection, sql, user.getPassword(), user.getEmail(), user.getAccount());
     }
 
     public List<User> findAll() {
