+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,63 +1,16 @@
 package com.techcourse.service;
 
-import com.interface21.dao.DataAccessException;
-import com.interface21.jdbc.datasource.DataSourceUtils;
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
-import java.sql.SQLException;
-import javax.sql.DataSource;
 
-public class UserService {
+public interface UserService {
 
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
+    User findById(final long id);
 
-    private final DataSource dataSource;
+    void save(final User user);
 
-    public UserService(
-            final UserDao userDao,
-            final UserHistoryDao userHistoryDao,
-            final DataSource dataSource
-    ) {
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-        this.dataSource = dataSource;
-    }
-
-    public User findById(final long id) {
-        return userDao.findById(id);
-    }
-
-    public void insert(final User user) {
-        userDao.insert(user);
-    }
-
-    public void changePassword(
+    void changePassword(
             final long id,
             final String newPassword,
-            final String createBy
-    ) {
-        final var connection = DataSourceUtils.getConnection(dataSource);
-        try {
-            connection.setAutoCommit(false);
-
-            try {
-                final var user = findById(id);
-                user.changePassword(newPassword);
-                userDao.update(connection, user);
-                userHistoryDao.log(connection, new UserHistory(user, createBy));
-
-                connection.commit();
-            } catch (final Exception e) {
-                connection.rollback();
-                throw new DataAccessException(e);
-            }
-        } catch (final SQLException e) {
-            throw new DataAccessException(e);
-        } finally {
-            DataSourceUtils.releaseConnection(connection, dataSource);
-        }
-    }
+            final String createdBy
+    );
 }
