+++ app/src/test/java/com/techcourse/service/MockUserHistoryDao.java
@@ -4,7 +4,6 @@ import com.interface21.dao.DataAccessException;
 import com.interface21.jdbc.core.JdbcTemplate;
 import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.UserHistory;
-import java.sql.Connection;
 
 public class MockUserHistoryDao extends UserHistoryDao {
 
@@ -13,10 +12,7 @@ public class MockUserHistoryDao extends UserHistoryDao {
     }
 
     @Override
-    public void log(
-            final Connection connection,
-            final UserHistory userHistory
-    ) {
+    public void log(final UserHistory userHistory) {
         throw new DataAccessException();
     }
 }
